[Laucnhpad] Extract CORS Filter as a Component
