+++ src/main/java/com/snapchat/launchpad/common/configs/SecurityConfig.java
@@ -1,6 +1,7 @@
 package com.snapchat.launchpad.common.configs;
 
 
+import com.snapchat.launchpad.common.components.CorsFilter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
@@ -19,23 +20,21 @@ import org.springframework.web.filter.OncePerRequestFilter;
 @EnableWebSecurity
 public class SecurityConfig {
 
+    private final CorsFilter corsFilter;
     private final OncePerRequestFilter launchpadSecurityFilter;
-
-    CorsFilter corsFilter() {
-        CorsFilter filter = new CorsFilter();
-        return filter;
-    }
 
     @Autowired
     public SecurityConfig(
-            @Qualifier("launchpadSecurityFilter") OncePerRequestFilter launchpadSecurityFilter) {
+            @Qualifier("launchpadSecurityFilter") OncePerRequestFilter launchpadSecurityFilter,
+            CorsFilter corsFilter) {
+        this.corsFilter = corsFilter;
         this.launchpadSecurityFilter = launchpadSecurityFilter;
     }
 
     @Bean
     public SecurityFilterChain configure(HttpSecurity http) throws Exception {
         // use custom filter for cors
-        http = http.addFilterBefore(corsFilter(), SessionManagementFilter.class);
+        http = http.addFilterBefore(corsFilter, SessionManagementFilter.class);
 
         // disable CSRF
         http = http.csrf().disable();
